requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.4.3
langchain==0.2.14
langchain-core==0.2.34
langchain-community==0.2.12
langchain-groq==0.1.4
langgraph==0.1.11
httpx==0.27.2
anyio==4.4.0

